Rename Chip.svelte to DeveloperIndicatorPill.svelte
