+++ packages/shared/components/atoms/pills/index.js
@@ -1,4 +1,4 @@
 export { default as ActivityStatusPill } from './ActivityStatusPill.svelte'
 export { default as ActivityAsyncStatusPill } from './ActivityAsyncStatusPill.svelte'
-export { default as Chip } from './Chip.svelte'
+export { default as DeveloperIndicatorPill } from './DeveloperIndicatorPill.svelte'
 export { default as Pill } from './Pill.svelte'
